--- src/components/Cards destacados y pintores/index.css
@@ -1,13 +1,0 @@
-body{
-  font-family: Arial, Helvetica, sans-serif;
-  background-color: #f4f4f9;
-  margin: 0;
-  padding: 0;
-}
-
-
-h2{
-  text-align:  center;
-  color: red;
-  margin-top: 20px;
-}
