+++ src/components/CardsDES_PIN/productCard.css
@@ -1,4 +1,5 @@
 .product-card {
+    
     border: 1px solid #ddd;
     border-radius: 8px;
     padding: 10px;
