+++ src/components/CardsDES_PIN/productList.css
@@ -9,3 +9,4 @@
     justify-content: flex-start; /* Alinea las tarjetas al inicio */
     align-items: flex-start; /* Alinea las tarjetas en la parte superior */
 }
+
