+++ src/components/Header/Header.css
@@ -3,6 +3,7 @@
     position: fixed;
     top: 0;
     z-index: 1000;
+    
   }
   
   .header-top {
