+++ src/pages/Marca.css
@@ -1,5 +1,5 @@
 .marcas-container {
-    padding-top: 120px; /* Espacio para el header fijo */
+    padding-top: 150px; /* Espacio para el header fijo */
     padding-bottom: 50px;
     background-color: #f5f5f5;
     min-height: 100vh;
